fix: Resolve compiler warnings across multiple modules

+++ JSON/src/Object.cpp
@@ -186,7 +186,7 @@ Object::NameList Object::getNames() const
 
 void Object::stringify(std::ostream& out, unsigned int indent, int step) const
 {
-	if (step < 0) step = indent;
+	if (step < 0) step = static_cast<int>(indent);
 
 	if (!_preserveInsOrder)
 		doStringify(_values, out, indent, step);
